apply review suggestions, rename file
